+++ src/AngleSharp/Io/DefaultHttpRequester.cs
@@ -1,6 +1,7 @@
 namespace AngleSharp.Io
 {
     using AngleSharp.Common;
+    using AngleSharp.Dom;
     using AngleSharp.Text;
     using System;
     using System.Collections.Generic;