+++ src/AngleSharp/Io/DefaultResponse.cs
@@ -2,6 +2,7 @@
 
 namespace AngleSharp.Io
 {
+    using AngleSharp.Dom;
     using System;
     using System.Collections.Generic;
     using System.IO;