+++ src/AngleSharp/Io/Download.cs
@@ -1,5 +1,6 @@
 namespace AngleSharp.Io
 {
+    using AngleSharp.Dom;
     using System;
     using System.Threading;
     using System.Threading.Tasks;