+++ src/AngleSharp/Io/VirtualResponse.cs
@@ -3,6 +3,7 @@
 namespace AngleSharp.Io
 {
     using AngleSharp.Common;
+    using AngleSharp.Dom;
     using AngleSharp.Text;
     using System;
     using System.Collections.Generic;