+++ src/AngleSharp/Text/CharExtensions.cs
@@ -16,10 +16,7 @@ namespace AngleSharp.Text
         /// <returns>
         /// The integer value or undefined behavior if invalid.
         /// </returns>
-        public static Int32 FromHex(this Char c)
-        {
-            return c.IsDigit() ? c - 0x30 : c - (c.IsLowercaseAscii() ? 0x57 : 0x37);
-        }
+        public static Int32 FromHex(this Char c) => c.IsDigit() ? c - 0x30 : c - (c.IsLowercaseAscii() ? 0x57 : 0x37);
 
         /// <summary>
         /// Transforms the given number to a hexadecimal string.
@@ -41,10 +38,7 @@ namespace AngleSharp.Text
         /// </summary>
         /// <param name="character">The single character.</param>
         /// <returns>A minimal digit lower case hexadecimal string.</returns>
-        public static String ToHex(this Char character)
-        {
-            return ((Int32)character).ToString("x");
-        }
+        public static String ToHex(this Char character) => ((Int32)character).ToString("x");
 
         /// <summary>
         /// Determines if the given character is in the given range.
@@ -54,10 +48,7 @@ namespace AngleSharp.Text
         /// <param name="upper">The upper bound of the range.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsInRange(this Char c, Int32 lower, Int32 upper)
-        {
-            return c >= lower && c <= upper;
-        }
+        public static Boolean IsInRange(this Char c, Int32 lower, Int32 upper) => c >= lower && c <= upper;
 
         /// <summary>
         /// Determines if the given character is allowed as-it-is in queries.
@@ -78,13 +69,11 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsNormalPathCharacter(this Char c)
-        {
-            return c.IsInRange(0x20, 0x7e) && c != Symbols.DoubleQuote &&
-                c != Symbols.CurvedQuote && c != Symbols.Num &&
-                c != Symbols.LessThan && c != Symbols.GreaterThan && 
-                c != Symbols.Space && c != Symbols.QuestionMark;
-        }
+        public static Boolean IsNormalPathCharacter(this Char c) =>
+            c.IsInRange(0x20, 0x7e) && c != Symbols.DoubleQuote &&
+            c != Symbols.CurvedQuote && c != Symbols.Num &&
+            c != Symbols.LessThan && c != Symbols.GreaterThan &&
+            c != Symbols.Space && c != Symbols.QuestionMark;
 
         /// <summary>
         /// Determines if the given character is a uppercase character (A-Z) as
@@ -94,10 +83,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsUppercaseAscii(this Char c)
-        {
-            return c >= 0x41 && c <= 0x5a;
-        }
+        public static Boolean IsUppercaseAscii(this Char c) => c >= 0x41 && c <= 0x5a;
 
         /// <summary>
         /// Determines if the given character is a lowercase character (a-z) as
@@ -107,10 +93,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsLowercaseAscii(this Char c)
-        {
-            return c >= 0x61 && c <= 0x7a;
-        }
+        public static Boolean IsLowercaseAscii(this Char c) => c >= 0x61 && c <= 0x7a;
 
         /// <summary>
         /// Determines if the given character is a alphanumeric character
@@ -120,10 +103,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsAlphanumericAscii(this Char c)
-        {
-            return c.IsDigit() || c.IsUppercaseAscii() || c.IsLowercaseAscii();
-        }
+        public static Boolean IsAlphanumericAscii(this Char c) => c.IsDigit() || c.IsUppercaseAscii() || c.IsLowercaseAscii();
 
         /// <summary>
         /// Determines if the given character is a hexadecimal (0-9a-fA-F) as
@@ -133,10 +113,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsHex(this Char c)
-        {
-            return c.IsDigit() || (c >= 0x41 && c <= 0x46) || (c >= 0x61 && c <= 0x66);
-        }
+        public static Boolean IsHex(this Char c) => c.IsDigit() || (c >= 0x41 && c <= 0x46) || (c >= 0x61 && c <= 0x66);
 
         /// <summary>
         /// Gets if the character is actually a non-ascii character.
@@ -144,10 +121,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsNonAscii(this Char c)
-        {
-            return c != Symbols.EndOfFile && c >= 0x80;
-        }
+        public static Boolean IsNonAscii(this Char c) => c != Symbols.EndOfFile && c >= 0x80;
 
         /// <summary>
         /// Gets if the character is actually a non-printable (special)
@@ -156,10 +130,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsNonPrintable(this Char c)
-        {
-            return (c >= 0x0 && c <= 0x8) || (c >= 0xe && c <= 0x1f) || (c >= 0x7f && c <= 0x9f);
-        }
+        public static Boolean IsNonPrintable(this Char c) => (c >= 0x0 && c <= 0x8) || (c >= 0xe && c <= 0x1f) || (c >= 0x7f && c <= 0x9f);
 
         /// <summary>
         /// Gets if the character is actually a (A-Z,a-z) letter.
@@ -167,10 +138,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsLetter(this Char c)
-        {
-            return IsUppercaseAscii(c) || IsLowercaseAscii(c);
-        }
+        public static Boolean IsLetter(this Char c) => IsUppercaseAscii(c) || IsLowercaseAscii(c);
 
         /// <summary>
         /// Gets if the character is actually a name character.
@@ -178,10 +146,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsName(this Char c)
-        {
-            return c.IsNonAscii() || c.IsLetter() || c == Symbols.Underscore || c == Symbols.Minus || c.IsDigit();
-        }
+        public static Boolean IsName(this Char c) => c.IsNonAscii() || c.IsLetter() || c == Symbols.Underscore || c == Symbols.Minus || c.IsDigit();
 
         /// <summary>
         /// Determines if the given character is a valid character for starting
@@ -190,10 +155,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsNameStart(this Char c)
-        {
-            return c.IsNonAscii() || c.IsUppercaseAscii() || c.IsLowercaseAscii() || c == Symbols.Underscore;
-        }
+        public static Boolean IsNameStart(this Char c) => c.IsNonAscii() || c.IsUppercaseAscii() || c.IsLowercaseAscii() || c == Symbols.Underscore;
 
         /// <summary>
         /// Determines if the given character is a line break character as
@@ -203,10 +165,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsLineBreak(this Char c)
-        {
-            return c == Symbols.LineFeed || c == Symbols.CarriageReturn;
-        }
+        public static Boolean IsLineBreak(this Char c) => c == Symbols.LineFeed || c == Symbols.CarriageReturn;
 
         /// <summary>
         /// Determines if the given character is a space character as specified
@@ -216,10 +175,8 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsSpaceCharacter(this Char c)
-        {
-            return c == Symbols.Space || c == Symbols.Tab || c == Symbols.LineFeed || c == Symbols.CarriageReturn || c == Symbols.FormFeed;
-        }
+        public static Boolean IsSpaceCharacter(this Char c) =>
+            c == Symbols.Space || c == Symbols.Tab || c == Symbols.LineFeed || c == Symbols.CarriageReturn || c == Symbols.FormFeed;
 
         /// <summary>
         /// Determines if the given character is a white-space character as
@@ -229,12 +186,10 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsWhiteSpaceCharacter(this Char c)
-        {
-            return c.IsInRange(0x0009, 0x000d) || c == 0x0020 || c == 0x0085 || c == 0x00a0 ||
-                    c == 0x1680 || c == 0x180e || c.IsInRange(0x2000, 0x200a) || c == 0x2028 ||
-                    c == 0x2029 || c == 0x202f || c == 0x205f || c == 0x3000;
-        }
+        public static Boolean IsWhiteSpaceCharacter(this Char c) =>
+            c.IsInRange(0x0009, 0x000d) || c == 0x0020 || c == 0x0085 || c == 0x00a0 ||
+            c == 0x1680 || c == 0x180e || c.IsInRange(0x2000, 0x200a) || c == 0x2028 ||
+            c == 0x2029 || c == 0x202f || c == 0x205f || c == 0x3000;
 
         /// <summary>
         /// Determines if the given character is a digit (0-9) as specified
@@ -244,10 +199,7 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsDigit(this Char c)
-        {
-            return c >= 0x30 && c <= 0x39;
-        }
+        public static Boolean IsDigit(this Char c) => c >= 0x30 && c <= 0x39;
 
         /// <summary>
         /// Determines if the given character is a valid url code point as specified here:
@@ -255,19 +207,17 @@ namespace AngleSharp.Text
         /// </summary>
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
-        public static Boolean IsUrlCodePoint(this Char c)
-        {
-            return c.IsAlphanumericAscii() || c == Symbols.ExclamationMark || c == Symbols.Dollar || c == Symbols.Ampersand ||
-                   c == Symbols.SingleQuote || c == Symbols.RoundBracketOpen || c == Symbols.RoundBracketClose ||
-                   c == Symbols.Asterisk || c == Symbols.Plus || c == Symbols.Minus || c == Symbols.Comma ||
-                   c == Symbols.Dot || c == Symbols.Solidus || c == Symbols.Colon || c == Symbols.Semicolon ||
-                   c == Symbols.Equality || c == Symbols.QuestionMark || c == Symbols.At || c == Symbols.Underscore ||
-                   c == Symbols.Tilde || c.IsInRange(0xa0, 0xd7ff) || c.IsInRange(0xe000, 0xfdcf) || c.IsInRange(0xfdf0, 0xfffd) ||
-                   c.IsInRange(0x10000, 0x1FFFD) || c.IsInRange(0x20000, 0x2fffd) || c.IsInRange(0x30000, 0x3fffd) || c.IsInRange(0x40000, 0x4fffd) ||
-                   c.IsInRange(0x50000, 0x5fffd) || c.IsInRange(0x60000, 0x6fffd) || c.IsInRange(0x70000, 0x7fffd) || c.IsInRange(0x80000, 0x8fffd) ||
-                   c.IsInRange(0x90000, 0x9fffd) || c.IsInRange(0xa0000, 0xafffd) || c.IsInRange(0xb0000, 0xbfffd) || c.IsInRange(0xc0000, 0xcfffd) ||
-                   c.IsInRange(0xd0000, 0xdfffd) || c.IsInRange(0xe0000, 0xefffd) || c.IsInRange(0xf0000, 0xffffd) || c.IsInRange(0x100000, 0x10fffd);
-        }
+        public static Boolean IsUrlCodePoint(this Char c) =>
+            c.IsAlphanumericAscii() || c == Symbols.ExclamationMark || c == Symbols.Dollar || c == Symbols.Ampersand ||
+            c == Symbols.SingleQuote || c == Symbols.RoundBracketOpen || c == Symbols.RoundBracketClose ||
+            c == Symbols.Asterisk || c == Symbols.Plus || c == Symbols.Minus || c == Symbols.Comma ||
+            c == Symbols.Dot || c == Symbols.Solidus || c == Symbols.Colon || c == Symbols.Semicolon ||
+            c == Symbols.Equality || c == Symbols.QuestionMark || c == Symbols.At || c == Symbols.Underscore ||
+            c == Symbols.Tilde || c.IsInRange(0xa0, 0xd7ff) || c.IsInRange(0xe000, 0xfdcf) || c.IsInRange(0xfdf0, 0xfffd) ||
+            c.IsInRange(0x10000, 0x1FFFD) || c.IsInRange(0x20000, 0x2fffd) || c.IsInRange(0x30000, 0x3fffd) || c.IsInRange(0x40000, 0x4fffd) ||
+            c.IsInRange(0x50000, 0x5fffd) || c.IsInRange(0x60000, 0x6fffd) || c.IsInRange(0x70000, 0x7fffd) || c.IsInRange(0x80000, 0x8fffd) ||
+            c.IsInRange(0x90000, 0x9fffd) || c.IsInRange(0xa0000, 0xafffd) || c.IsInRange(0xb0000, 0xbfffd) || c.IsInRange(0xc0000, 0xcfffd) ||
+            c.IsInRange(0xd0000, 0xdfffd) || c.IsInRange(0xe0000, 0xefffd) || c.IsInRange(0xf0000, 0xffffd) || c.IsInRange(0x100000, 0x10fffd);
 
         /// <summary>
         /// Determines if the given character is invalid, i.e. zero, above the
@@ -276,54 +226,6 @@ namespace AngleSharp.Text
         /// <param name="c">The character to examine.</param>
         /// <returns>The result of the test.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsInvalid(this Int32 c)
-        {
-            return c == 0 || c > Symbols.MaximumCodepoint || (c > 0xD800 && c < 0xDFFF);
-        }
-
-        /// <summary>
-        /// Determines if the given character is one of the two others.
-        /// </summary>
-        /// <param name="c">The character to test.</param>
-        /// <param name="a">The first option.</param>
-        /// <param name="b">The second option.</param>
-        /// <returns>The result of the test.</returns>
-        [Obsolete("Use pattern matching instead. This will be removed from a future release.")]
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsOneOf(this Char c, Char a, Char b)
-        {
-            return a == c || b == c;
-        }
-
-        /// <summary>
-        /// Determines if the given character is one of the three others.
-        /// </summary>
-        /// <param name="c">The character to test.</param>
-        /// <param name="o1">The first option.</param>
-        /// <param name="o2">The second option.</param>
-        /// <param name="o3">The third option.</param>
-        /// <returns>The result of the test.</returns>
-        [Obsolete("Use pattern matching instead. This will be removed from a future release.")]
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsOneOf(this Char c, Char o1, Char o2, Char o3)
-        {
-            return c == o1 || c == o2 || c == o3;
-        }
-
-        /// <summary>
-        /// Determines if the given character is one of the four others.
-        /// </summary>
-        /// <param name="c">The character to test.</param>
-        /// <param name="o1">The first option.</param>
-        /// <param name="o2">The second option.</param>
-        /// <param name="o3">The third option.</param>
-        /// <param name="o4">The fourth option.</param>
-        /// <returns>The result of the test.</returns>
-        [Obsolete("Use pattern matching instead. This will be removed from a future release.")]
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Boolean IsOneOf(this Char c, Char o1, Char o2, Char o3, Char o4)
-        {
-            return c == o1 || c == o2 || c == o3 || c == o4;
-        }
+        public static Boolean IsInvalid(this Int32 c) => c == 0 || c > Symbols.MaximumCodepoint || (c > 0xD800 && c < 0xDFFF);
     }
 }